+++ scripts/run_GWAS.py
@@ -275,6 +275,7 @@ def main():
         alpha=args.alpha,
         n_eff=args.n_eff,
         max_genotype_dosage=args.max_genotype_dosage,
+        min_mac=args.min_mac,
         farmcpu_params=farmcpu_params,
         bayesloco_params=(bayesloco_params or None),
         include_standard_errors=args.include_standard_errors,